Improve naming and structure
